backend/requirements.txt
fastapi==0.95.0
uvicorn==0.22.0
pydantic==1.10.2
pandas==1.5.3
PyYAML==6.0
